+++ eth_defi/hypersync/server.py
@@ -6,7 +6,7 @@
 from web3 import Web3
 
 #: HyperSync server URLs by chain id.
-#: Updated 2026-07-09.
+#: Updated 2026-07-10.
 HYPERSYNC_SERVES = {
     1: "https://eth.hypersync.xyz",  # Ethereum Mainnet
     10: "https://optimism.hypersync.xyz",  # Optimism
@@ -46,6 +46,7 @@ HYPERSYNC_SERVES = {
     4114: "https://citrea.hypersync.xyz",  # Citrea
     4200: "https://merlin.hypersync.xyz",  # Merlin
     4201: "https://lukso-testnet.hypersync.xyz",  # Lukso Testnet
+    4217: "https://tempo.hypersync.xyz",  # Tempo
     4326: "https://megaeth.hypersync.xyz",  # MegaETH
     4663: "https://4663.hypersync.xyz",  # Robinhood Chain
     5000: "https://mantle.hypersync.xyz",  # Mantle